extracted containers from components
