+++ src/containers/styles/Details.css
