+++ src/index.jsx
@@ -6,19 +6,19 @@ import { Router, Route, Switch } from 'react-router-dom';
 import promise from 'redux-promise';
 
 import App from './components/app';
-import Home from './components/home';
-import Login from './components/login';
-import CreateProject from './components/create_project';
-import UpdateProject from './components/update_project';
-import UpdateUser from './components/update_user';
+import Home from './containers/home';
+import Login from './containers/login';
+import CreateProject from './containers/create_project';
+import UpdateProject from './containers/update_project';
+import UpdateUser from './containers/update_user';
 import NotFound from './components/not_found';
-import Project from './components/project';
-import Projects from './components/projects';
-import RequireAuth from './components/authentication';
-import RspecReports from './components/rspec_reports';
-import CreateUser from './components/create_user';
-import User from './components/user';
-import Users from './components/users';
+import Project from './containers/project';
+import Projects from './containers/projects';
+import RequireAuth from './containers/authentication';
+import RspecReports from './containers/rspec_reports';
+import CreateUser from './containers/create_user';
+import User from './containers/user';
+import Users from './containers/users';
 import history from './history';
 import reducers from './reducers';
